Change: Slight refactor for more clear struct.
